@@ -21,44 +20,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "16.2.4",
-    "@angular/common": "16.2.4",
-    "@angular/compiler": "16.2.4",
-    "@angular/core": "16.2.4",
-    "@angular/forms": "16.2.4",
-    "@angular/platform-browser": "16.2.4",
-    "@angular/platform-browser-dynamic": "16.2.4",
-    "@angular/platform-server": "16.2.4",
-    "@angular/router": "16.2.4",
-    "@delon/abc": "^16.3.0",
-    "@delon/acl": "^16.3.0",
-    "@delon/auth": "^16.3.0",
-    "@delon/cache": "^16.3.0",
-    "@delon/chart": "^16.3.0",
-    "@delon/form": "^16.3.0",
-    "@delon/mock": "^16.3.0",
-    "@delon/theme": "^16.3.0",
-    "@delon/util": "^16.3.0",
-    "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^4.4.1",
+    "@angular/animations": "17.0.8",
+    "@angular/common": "17.0.8",
+    "@angular/compiler": "17.0.8",
+    "@angular/core": "17.0.8",
+    "@angular/forms": "17.0.8",
+    "@angular/platform-browser": "17.0.8",
+    "@angular/platform-browser-dynamic": "17.0.8",
+    "@angular/platform-server": "17.0.8",
+    "@angular/router": "17.0.8",
+    "@delon/abc": "^17.0.5",
+    "@delon/acl": "^17.0.5",
+    "@delon/auth": "^17.0.5",
+    "@delon/cache": "^17.0.5",
+    "@delon/chart": "^17.0.5",
+    "@delon/form": "^17.0.5",
+    "@delon/mock": "^17.0.5",
+    "@delon/theme": "^17.0.5",
+    "@delon/util": "^17.0.5",
+    "bootstrap": "^5.3.2",
     "date-fns": "^2.12.0",
-    "ng-zorro-antd": "^16.1.0",
+    "ng-zorro-antd": "^17.1.0",
     "rxjs": "~7.8.0",
     "screenfull": "^5.0.2",
     "tslib": "^2.0.0",
     "wangeditor": "^4.7.4",
-    "zone.js": "~0.13.0"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.1",
-    "@angular-eslint/builder": "~16.1.1",
-    "@angular-eslint/eslint-plugin": "~16.1.1",
-    "@angular-eslint/eslint-plugin-template": "~16.1.1",
-    "@angular-eslint/schematics": "~16.1.1",
-    "@angular-eslint/template-parser": "~16.1.1",
-    "@angular/cli": "^16.2.1",
-    "@angular/compiler-cli": "16.2.4",
-    "@angular/language-service": "^16.2.0",
+    "@angular-devkit/build-angular": "^17.0.8",
+    "@angular-eslint/builder": "~17.1.1",
+    "@angular-eslint/eslint-plugin": "~17.1.1",
+    "@angular-eslint/eslint-plugin-template": "~17.1.1",
+    "@angular-eslint/schematics": "~17.1.1",
+    "@angular-eslint/template-parser": "~17.1.1",
+    "@angular/cli": "^17.0.8",
+    "@angular/compiler-cli": "17.0.8",
+    "@angular/language-service": "^17.0.8",
     "@delon/testing": "^16.3.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
@@ -79,17 +76,14 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-alain": "^16.3.0",
+    "ng-alain": "^17.0.5",
     "prettier": "^3.0.2",
     "source-map-explorer": "^2.5.3",
-    "typescript": "~4.9.5"
+    "typescript": "~5.2.2"
   },
   "optionalDependencies": {
     "protractor": "~7.0.0",
     "ts-node": "~8.4.1",
     "tslint": "~6.1.0"
-  },
-  "husky": {
-    "hooks": {}
   }
 }
