@@ -1,7 +1,24 @@
-import { Component } from '@angular/core';
-import { SettingsService, User } from '@delon/theme';
-import { LayoutDefaultOptions } from '@delon/theme/layout-default';
+import { Component, inject } from '@angular/core';
+import { RouterLink, RouterOutlet } from '@angular/router';
+import { I18nPipe, SettingsService, User } from '@delon/theme';
+import { LayoutDefaultModule, LayoutDefaultOptions } from '@delon/theme/layout-default';
+import { SettingDrawerModule } from '@delon/theme/setting-drawer';
+import { ThemeBtnComponent } from '@delon/theme/theme-btn';
 import { environment } from '@env/environment';
+import { NzAvatarModule } from 'ng-zorro-antd/avatar';
+import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
+import { NzIconModule } from 'ng-zorro-antd/icon';
+import { NzMenuModule } from 'ng-zorro-antd/menu';
+
+import { HeaderClearStorageComponent } from './widgets/clear-storage.component';
+import { HeaderFullScreenComponent } from './widgets/fullscreen.component';
+import { HeaderI18nComponent } from './widgets/i18n.component';
+import { HeaderIconComponent } from './widgets/icon.component';
+import { HeaderNotifyComponent } from './widgets/notify.component';
+import { HeaderRTLComponent } from './widgets/rtl.component';
+import { HeaderSearchComponent } from './widgets/search.component';
+import { HeaderTaskComponent } from './widgets/task.component';
+import { HeaderUserComponent } from './widgets/user.component';
 
 @Component({
   selector: 'layout-basic',
@@ -32,7 +49,7 @@ import { environment } from '@env/environment';
       </layout-default-header-item>
        -->
       <layout-default-header-item direction="right">
-        <header-notify></header-notify>
+        <header-notify/>
       </layout-default-header-item>
       <!--
       <layout-default-header-item direction="right" hidden="mobile">
@@ -45,7 +62,8 @@ import { environment } from '@env/environment';
       </layout-default-header-item>
       -->
       <layout-default-header-item direction="right" hidden="mobile">
-        <div layout-default-header-item-trigger nz-dropdown [nzDropdownMenu]="settingsMenu" nzTrigger="click" nzPlacement="bottomRight">
+        <div layout-default-header-item-trigger nz-dropdown [nzDropdownMenu]="settingsMenu" nzTrigger="click"
+             nzPlacement="bottomRight">
           <i nz-icon nzType="setting"></i>
         </div>
         <nz-dropdown-menu #settingsMenu="nzDropdownMenu">
@@ -82,10 +100,12 @@ import { environment } from '@env/environment';
           <ul nz-menu>
             <li nz-menu-item routerLink="/pro/account/center">
               <i nz-icon nzType="user" class="mr-sm"></i>
-              个人中心</li>
+              个人中心
+            </li>
             <li nz-menu-item routerLink="/pro/account/settings">
               <i nz-icon nzType="setting" class="mr-sm"></i>
-              个人设置</li>
+              个人设置
+            </li>
           </ul>
         </nz-dropdown-menu>
       </ng-template>
@@ -94,20 +114,47 @@ import { environment } from '@env/environment';
       </ng-template>
     </layout-default>
 
-    <setting-drawer *ngIf="showSettingDrawer"></setting-drawer>
+    <!--
+    @if (showSettingDrawer) {
+      <setting-drawer></setting-drawer>
+    }
     <theme-btn></theme-btn>
+    -->
   `,
+  standalone: true,
+  imports: [
+    RouterOutlet,
+    RouterLink,
+    I18nPipe,
+    LayoutDefaultModule,
+    NzIconModule,
+    NzMenuModule,
+    NzDropDownModule,
+    NzAvatarModule,
+    SettingDrawerModule,
+    ThemeBtnComponent,
+    HeaderSearchComponent,
+    HeaderNotifyComponent,
+    HeaderTaskComponent,
+    HeaderIconComponent,
+    HeaderRTLComponent,
+    HeaderI18nComponent,
+    HeaderClearStorageComponent,
+    HeaderFullScreenComponent,
+    HeaderUserComponent
+  ]
 })
 export class LayoutBasicComponent {
+  private readonly settings = inject(SettingsService);
   options: LayoutDefaultOptions = {
     logoExpanded: `./assets/logo-full.svg`,
     logoCollapsed: `./assets/logo.svg`,
   };
   searchToggleStatus = false;
   showSettingDrawer = !environment.production;
+
   get user(): User {
     return this.settings.user;
   }
 
-  constructor(private settings: SettingsService) {}
 }
