@@ -1,4 +1,6 @@
-import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
+import { ChangeDetectionStrategy, Component, HostListener, inject } from '@angular/core';
+import { I18nPipe } from '@delon/theme';
+import { NzIconModule } from 'ng-zorro-antd/icon';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { NzModalService } from 'ng-zorro-antd/modal';
 
@@ -11,13 +13,13 @@ import { NzModalService } from 'ng-zorro-antd/modal';
   host: {
     '[class.flex-1]': 'true'
   },
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [NzIconModule, I18nPipe]
 })
 export class HeaderClearStorageComponent {
-  constructor(
-    private modalSrv: NzModalService,
-    private messageSrv: NzMessageService
-  ) {}
+  private readonly modalSrv = inject(NzModalService);
+  private readonly messageSrv = inject(NzMessageService);
 
   @HostListener('click')
   _click(): void {
