@@ -1,36 +1,47 @@
 import { DOCUMENT } from '@angular/common';
-import { ChangeDetectionStrategy, Component, Inject, Input } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input, booleanAttribute, inject } from '@angular/core';
 import { I18NService } from '@core';
-import { ALAIN_I18N_TOKEN, SettingsService } from '@delon/theme';
-import { BooleanInput, InputBoolean } from '@delon/util/decorator';
+import { ALAIN_I18N_TOKEN, I18nPipe, SettingsService } from '@delon/theme';
+import { NzDropDownModule } from 'ng-zorro-antd/dropdown';
+import { NzIconModule } from 'ng-zorro-antd/icon';
+import { NzMenuModule } from 'ng-zorro-antd/menu';
 
 @Component({
   selector: 'header-i18n',
   template: `
-    <div *ngIf="showLangText" nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight">
-      <i nz-icon nzType="global"></i>
-      {{ 'menu.lang' | i18n }}
-      <i nz-icon nzType="down"></i>
-    </div>
-    <i *ngIf="!showLangText" nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight" nz-icon nzType="global"></i>
+    @if (showLangText) {
+      <div nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight">
+        <i nz-icon nzType="global"></i>
+        {{ 'menu.lang' | i18n }}
+        <i nz-icon nzType="down"></i>
+      </div>
+    } @else {
+      <i nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight" nz-icon nzType="global"></i>
+    }
     <nz-dropdown-menu #langMenu="nzDropdownMenu">
       <ul nz-menu>
-        <li nz-menu-item *ngFor="let item of langs" [nzSelected]="item.code === curLangCode" (click)="change(item.code)">
-          <span role="img" [attr.aria-label]="item.text" class="pr-xs">{{ item.abbr }}</span>
-          {{ item.text }}
-        </li>
+        @for (item of langs; track $index) {
+          <li nz-menu-item [nzSelected]="item.code === curLangCode" (click)="change(item.code)">
+            <span role="img" [attr.aria-label]="item.text" class="pr-xs">{{ item.abbr }}</span>
+            {{ item.text }}
+          </li>
+        }
       </ul>
     </nz-dropdown-menu>
   `,
   host: {
     '[class.flex-1]': 'true'
   },
-  changeDetection: ChangeDetectionStrategy.OnPush
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [I18nPipe, NzDropDownModule, NzIconModule, NzMenuModule]
 })
 export class HeaderI18nComponent {
-  static ngAcceptInputType_showLangText: BooleanInput;
+  private readonly settings = inject(SettingsService);
+  private readonly i18n = inject<I18NService>(ALAIN_I18N_TOKEN);
+  private readonly doc = inject(DOCUMENT);
   /** Whether to display language text */
-  @Input() @InputBoolean() showLangText = true;
+  @Input({ transform: booleanAttribute }) showLangText = true;
 
   get langs(): Array<{ code: string; text: string; abbr: string }> {
     return this.i18n.getLangs();
@@ -39,12 +50,6 @@ export class HeaderI18nComponent {
   get curLangCode(): string {
     return this.settings.layout.lang;
   }
-
-  constructor(
-    private settings: SettingsService,
-    @Inject(ALAIN_I18N_TOKEN) private i18n: I18NService,
-    @Inject(DOCUMENT) private doc: any
-  ) {}
 
   change(lang: string): void {
     const spinEl = this.doc.createElement('div');
