@@ -1,5 +1,7 @@
-import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
+import { UpperCasePipe } from '@angular/common';
+import { ChangeDetectionStrategy, Component, HostListener, inject } from '@angular/core';
 import { RTLService } from '@delon/theme';
+import { NzIconModule } from 'ng-zorro-antd/icon';
 
 @Component({
   selector: 'header-rtl',
@@ -7,14 +9,15 @@ import { RTLService } from '@delon/theme';
     <i nz-icon [nzType]="rtl.nextDir === 'rtl' ? 'border-left' : 'border-right'"></i>
     {{ rtl.nextDir | uppercase }}
   `,
-  // tslint:disable-next-line: no-host-metadata-property
   host: {
-    '[class.d-block]': 'true',
+    '[class.flex-1]': 'true'
   },
   changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [NzIconModule, UpperCasePipe]
 })
 export class HeaderRTLComponent {
-  constructor(public rtl: RTLService) {}
+  readonly rtl = inject(RTLService);
 
   @HostListener('click')
   toggleDirection(): void {
