@@ -1,3 +1,4 @@
+// 请参考：https://ng-alain.com/docs/i18n
 import { Platform } from '@angular/cdk/platform';
 import { registerLocaleData } from '@angular/common';
 import ngEn from '@angular/common/locales/en';
@@ -56,12 +57,12 @@ const LANGS: { [key: string]: LangConfigData } = {
   }
 };
 
-@Injectable({providedIn: 'root'})
+@Injectable({ providedIn: 'root' })
 export class I18NService extends AlainI18nBaseService {
   protected override _defaultLang = DEFAULT;
   private _langs = Object.keys(LANGS).map(code => {
     const item = LANGS[code];
-    return {code, text: item.text, abbr: item.abbr};
+    return { code, text: item.text, abbr: item.abbr };
   });
 
   constructor(
