@@ -1,11 +1,13 @@
 import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   selector: 'layout-blank',
-  template: `<router-outlet></router-outlet> `,
-  // tslint:disable-next-line: no-host-metadata-property
+  template: `<router-outlet />`,
   host: {
-    '[class.alain-blank]': 'true',
+    '[class.alain-blank]': 'true'
   },
+  standalone: true,
+  imports: [RouterOutlet]
 })
 export class LayoutBlankComponent {}
