@@ -1,12 +1,22 @@
-import { Component, Inject, OnInit } from '@angular/core';
-import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
+import { Component, OnInit, inject } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import { GlobalFooterModule } from '@delon/abc/global-footer';
+import { DA_SERVICE_TOKEN } from '@delon/auth';
+import { ThemeBtnComponent } from '@delon/theme/theme-btn';
+import { NzIconModule } from 'ng-zorro-antd/icon';
+
+import { HeaderI18nComponent } from '../basic/widgets/i18n.component';
 
 @Component({
   selector: 'layout-passport',
   templateUrl: './passport.component.html',
   styleUrls: ['./passport.component.less'],
+  standalone: true,
+  imports: [RouterOutlet, HeaderI18nComponent, GlobalFooterModule, NzIconModule, ThemeBtnComponent]
 })
 export class LayoutPassportComponent implements OnInit {
+  private tokenService = inject(DA_SERVICE_TOKEN);
+
   links = [
     {
       title: '帮助',
@@ -21,8 +31,6 @@ export class LayoutPassportComponent implements OnInit {
       href: '',
     },
   ];
-
-  constructor(@Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService) {}
 
   ngOnInit(): void {
     this.tokenService.clear();
