@@ -1,40 +1,44 @@
-import { Router } from '@angular/router';
 import { Component, Inject } from '@angular/core';
-import { UntypedFormGroup, UntypedFormBuilder, Validators } from '@angular/forms';
-import { SettingsService } from '@delon/theme';
+import { FormControl, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
+import { I18nPipe, SettingsService, User } from '@delon/theme';
+import { NzAvatarModule } from 'ng-zorro-antd/avatar';
+import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzFormModule } from 'ng-zorro-antd/form';
+import { NzGridModule } from 'ng-zorro-antd/grid';
+import { NzInputModule } from 'ng-zorro-antd/input';
 
 @Component({
   selector: 'passport-lock',
   templateUrl: './lock.component.html',
   styleUrls: ['./lock.component.less'],
+  standalone: true,
+  imports: [ReactiveFormsModule, I18nPipe, NzAvatarModule, NzFormModule, NzGridModule, NzButtonModule, NzInputModule]
 })
 export class UserLockComponent {
-  f: UntypedFormGroup;
+  f = new FormGroup({
+    password: new FormControl('', { nonNullable: true, validators: [Validators.required] })
+  });
 
-  constructor(
-    fb: UntypedFormBuilder,
-    @Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
-    public settings: SettingsService,
-    private router: Router,
-  ) {
-    tokenService.clear();
-    this.f = fb.group({
-      password: [null, Validators.required],
-    });
+  get user(): User {
+    return this.settings.user;
   }
 
-  submit() {
-    // tslint:disable-next-line:forin
-    for (const i in this.f.controls) {
-      this.f.controls[i].markAsDirty();
-      this.f.controls[i].updateValueAndValidity();
-    }
+  constructor(
+    @Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
+    private settings: SettingsService,
+    private router: Router
+  ) {}
+
+  submit(): void {
+    this.f.controls.password.markAsDirty();
+    this.f.controls.password.updateValueAndValidity();
     if (this.f.valid) {
       console.log('Valid!');
       console.log(this.f.value);
       this.tokenService.set({
-        token: '123',
+        token: '123'
       });
       this.router.navigate(['dashboard']);
     }
