@@ -1,54 +1,74 @@
-import { Component, OnDestroy } from '@angular/core';
-import { Router } from '@angular/router';
-import { FormGroup, FormBuilder, Validators, FormControl } from '@angular/forms';
-import { NzMessageService } from 'ng-zorro-antd';
-import { _HttpClient } from '@delon/theme';
+import { HttpContext } from '@angular/common/http';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy } from '@angular/core';
+import { AbstractControl, FormBuilder, FormControl, ReactiveFormsModule, Validators } from '@angular/forms';
+import { Router, RouterLink } from '@angular/router';
+import { ALLOW_ANONYMOUS } from '@delon/auth';
+import { I18nPipe, _HttpClient } from '@delon/theme';
+import { MatchControl } from '@delon/util/form';
+import { NzAlertModule } from 'ng-zorro-antd/alert';
+import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzSafeAny } from 'ng-zorro-antd/core/types';
+import { NzFormModule } from 'ng-zorro-antd/form';
+import { NzGridModule } from 'ng-zorro-antd/grid';
+import { NzInputModule } from 'ng-zorro-antd/input';
+import { NzPopoverModule } from 'ng-zorro-antd/popover';
+import { NzProgressModule } from 'ng-zorro-antd/progress';
+import { NzSelectModule } from 'ng-zorro-antd/select';
+import { finalize } from 'rxjs';
 
 @Component({
   selector: 'passport-register',
   templateUrl: './register.component.html',
   styleUrls: ['./register.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [
+    ReactiveFormsModule,
+    I18nPipe,
+    RouterLink,
+    NzAlertModule,
+    NzFormModule,
+    NzInputModule,
+    NzPopoverModule,
+    NzProgressModule,
+    NzSelectModule,
+    NzGridModule,
+    NzButtonModule
+  ]
 })
 export class UserRegisterComponent implements OnDestroy {
-
-  constructor(fb: FormBuilder, private router: Router, public http: _HttpClient, public msg: NzMessageService) {
-    this.form = fb.group({
-      mail: [null, [Validators.required, Validators.email]],
-      password: [null, [Validators.required, Validators.minLength(6), UserRegisterComponent.checkPassword.bind(this)]],
-      confirm: [null, [Validators.required, Validators.minLength(6), UserRegisterComponent.passwordEquar]],
-      mobilePrefix: ['+86'],
-      mobile: [null, [Validators.required, Validators.pattern(/^1\d{10}$/)]],
-      captcha: [null, [Validators.required]],
-    });
-  }
+  constructor(
+    private fb: FormBuilder,
+    private router: Router,
+    private http: _HttpClient,
+    private cdr: ChangeDetectorRef
+  ) {}
 
   // #region fields
 
-  get mail() {
-    return this.form.controls.mail;
-  }
-  get password() {
-    return this.form.controls.password;
-  }
-  get confirm() {
-    return this.form.controls.confirm;
-  }
-  get mobile() {
-    return this.form.controls.mobile;
-  }
-  get captcha() {
-    return this.form.controls.captcha;
-  }
-  form: FormGroup;
+  form = this.fb.nonNullable.group(
+    {
+      mail: ['', [Validators.required, Validators.email]],
+      password: ['', [Validators.required, Validators.minLength(6), UserRegisterComponent.checkPassword.bind(this)]],
+      confirm: ['', [Validators.required, Validators.minLength(6)]],
+      mobilePrefix: ['+86'],
+      mobile: ['', [Validators.required, Validators.pattern(/^1\d{10}$/)]],
+      captcha: ['', [Validators.required]]
+    },
+    {
+      validators: MatchControl('password', 'confirm')
+    }
+  );
   error = '';
   type = 0;
+  loading = false;
   visible = false;
   status = 'pool';
   progress = 0;
-  passwordProgressMap = {
+  passwordProgressMap: { [key: string]: 'success' | 'normal' | 'exception' } = {
     ok: 'success',
     pass: 'normal',
-    pool: 'exception',
+    pool: 'exception'
   };
 
   // #endregion
@@ -56,11 +76,14 @@ export class UserRegisterComponent implements OnDestroy {
   // #region get captcha
 
   count = 0;
-  interval$: any;
+  interval$: NzSafeAny;
 
-  static checkPassword(control: FormControl) {
-    if (!control) return null;
-    const self: any = this;
+  static checkPassword(control: FormControl): NzSafeAny {
+    if (!control) {
+      return null;
+    }
+    // eslint-disable-next-line @typescript-eslint/no-this-alias
+    const self: NzSafeAny = this;
     self.visible = !!control.value;
     if (control.value && control.value.length > 9) {
       self.status = 'ok';
@@ -75,50 +98,58 @@ export class UserRegisterComponent implements OnDestroy {
     }
   }
 
-  static passwordEquar(control: FormControl) {
-    if (!control || !control.parent) {
-      return null;
-    }
-    if (control.value !== control.parent.get('password')!.value) {
-      return { equar: true };
-    }
-    return null;
-  }
-
-  getCaptcha() {
-    if (this.mobile.invalid) {
-      this.mobile.markAsDirty({ onlySelf: true });
-      this.mobile.updateValueAndValidity({ onlySelf: true });
+  getCaptcha(): void {
+    const { mobile } = this.form.controls;
+    if (mobile.invalid) {
+      mobile.markAsDirty({ onlySelf: true });
+      mobile.updateValueAndValidity({ onlySelf: true });
       return;
     }
     this.count = 59;
+    this.cdr.detectChanges();
     this.interval$ = setInterval(() => {
       this.count -= 1;
-      if (this.count <= 0) clearInterval(this.interval$);
+      this.cdr.detectChanges();
+      if (this.count <= 0) {
+        clearInterval(this.interval$);
+      }
     }, 1000);
   }
 
   // #endregion
 
-  submit() {
+  submit(): void {
     this.error = '';
     Object.keys(this.form.controls).forEach(key => {
-      this.form.controls[key].markAsDirty();
-      this.form.controls[key].updateValueAndValidity();
+      const control = (this.form.controls as NzSafeAny)[key] as AbstractControl;
+      control.markAsDirty();
+      control.updateValueAndValidity();
     });
     if (this.form.invalid) {
       return;
     }
 
     const data = this.form.value;
-    this.http.post('/register', data).subscribe(() => {
-      this.router.navigateByUrl('/passport/register-result', {
-        queryParams: { email: data.mail },
+    this.loading = true;
+    this.cdr.detectChanges();
+    this.http
+      .post('/register', data, null, {
+        context: new HttpContext().set(ALLOW_ANONYMOUS, true)
+      })
+      .pipe(
+        finalize(() => {
+          this.loading = false;
+          this.cdr.detectChanges();
+        })
+      )
+      .subscribe(() => {
+        this.router.navigate(['passport', 'register-result'], { queryParams: { email: data.mail } });
       });
-    });
   }
 
   ngOnDestroy(): void {
-    if (this.interval$) clearInterval(this.interval$);
+    if (this.interval$) {
+      clearInterval(this.interval$);
+    }
   }
 }
