@@ -5,4 +5,6 @@ export * from './utils/yuan';
 
 // Module
 export * from './shared.module';
-export * from './json-schema/json-schema.module';
+export * from './cell-widget/index'
+export * from './json-schema/index';
+export * from './st-widget/index';
