@@ -2,7 +2,7 @@
 // `ng build ---prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-import { DelonMockModule } from '@delon/mock';
+import { mockInterceptor, provideMockConfig } from '@delon/mock';
 import { Environment } from '@delon/theme';
 import * as MOCKDATA from '../../_mock';
 
@@ -12,15 +12,8 @@ export const environment = {
   api: {
     baseUrl: './',
     refreshTokenEnabled: true,
-    refreshTokenType: 'auth-refresh',
+    refreshTokenType: 'auth-refresh'
   },
-  modules: [DelonMockModule.forRoot({ data: MOCKDATA })],
+  providers: [provideMockConfig({ data: MOCKDATA })],
+  interceptorFns: [mockInterceptor]
 } as Environment;
-
-/*
- * In development mode, to ignore zone related error stack frames such as
- * `zone.run`, `zoneDelegate.invokeTask` for easier debugging, you can
- * import the following file, but please comment it out in production mode
- * because it will have performance impact when throw error
- */
-// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
