@@ -1,29 +1,6 @@
-import { enableProdMode, ViewEncapsulation } from '@angular/core';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { bootstrapApplication } from '@angular/platform-browser';
 
-import { AppModule } from './app/app.module';
-import { environment } from '@env/environment';
+import { AppComponent } from './app/app.component';
+import { appConfig } from './app/app.config';
 
-import { preloaderFinished } from '@delon/theme';
-preloaderFinished();
-
-import { NzSafeAny } from "ng-zorro-antd/core/types";
-
-if (environment.production) {
-  enableProdMode();
-}
-
-
-platformBrowserDynamic()
-  .bootstrapModule(AppModule, {
-    defaultEncapsulation: ViewEncapsulation.Emulated,
-    preserveWhitespaces: false,
-  })
-  .then((res) => {
-    const win = window as NzSafeAny;
-    if (win && win.appBootstrap) {
-      win.appBootstrap();
-    }
-    return res;
-  })
-  .catch((err) => console.error(err));
+bootstrapApplication(AppComponent, appConfig).catch(err => console.error(err));
